[INFRA] move pretty printing of chars from stream to core
